refactor(ui): fold has-attachments filter into the sort dropdown (#285)
Consolidate the entry-list header control: instead of a separate
FilterDropdown beside the sort dropdown, the "Has attachments" toggle
now lives at the bottom of the existing sort dropdown (under a Filter
section), so the header has a single dropdown.

- Add the has-attachments checkbox + handler to SortDropdown
- Remove FilterDropdown and its usages (desktop + mobile)
- Move toggle coverage into sort-dropdown.test.tsx
- Drop the now-unused entries.filter.filterEntries i18n key
