+++ src/locales/de/common.json
@@ -250,7 +250,6 @@
     "noEntriesMatchFilters": "Keine Einträge entsprechen den aktiven Filtern.",
     "filter": {
       "title": "Filter",
-      "filterEntries": "Einträge filtern",
       "hasAttachments": "Hat Anhänge"
     },
     "newEntry": "Neuer Eintrag",
